create a menu and debug

+++ src/user_delete.py
@@ -10,8 +10,10 @@ from read_inventory import Inventories
 from read_customer import Customers
 
 from datetime import datetime 
-from user_handle_message import send_long_text, cancel, build_date_keyboard, build_menu
+from user_handle_message import send_long_text, cancel, reply_message
 from helper import check_shipping_date
+
+from user_get_button_menu import get_customer_buttons, get_product_buttons, build_date_keyboard, get_supplier_buttons
 
 # Telegram Bot Token
 TOKEN = '6487583852:AAGH3YlPRpfuOtLt-GlWvyI4Ss-B1lxOdtA'
@@ -32,7 +34,8 @@ SALE_DELETE = range(2)
 
 # Function to start deleting a sale record by ID
 def start_delete_sale(update: Update, context: CallbackContext) -> int:
-    update.message.reply_text('请输入要删除的销售记录ID:')
+    message = '请输入要删除的销售记录ID:'
+    reply_message(update, context, message, None)
     return SALE_DELETE
 
 def delete_sale(update: Update, context: CallbackContext) -> int:
@@ -50,15 +53,11 @@ def delete_sale(update: Update, context: CallbackContext) -> int:
 CUSTOMER_DELETE_NAME, CUSTOMER_DELETE_CONFIRMATION = range(2)
 
 def customer_delete_start(update: Update, context: CallbackContext) -> int:
-    # 假设有一个方法返回所有客户名列表
-    customer_list = customers_manager.get_customer_name_list()
-    if len(customer_list) == 0:
-        update.message.reply_text('没有可以删除的客户名')
+    flag, reply_markup = get_customer_buttons(update, context, False, False)
+    if flag is False:
         return ConversationHandler.END
-    
-    keyboard = [[InlineKeyboardButton(customer, callback_data=customer)] for customer in customer_list]
-    reply_markup = InlineKeyboardMarkup(keyboard)
-    update.message.reply_text('请选择要删除的客户名：', reply_markup=reply_markup)
+    message = '请选择要删除的客户名：'
+    reply_message(update, context, message, reply_markup)
     return CUSTOMER_DELETE_NAME
 
 def customer_delete_name(update: Update, context: CallbackContext) -> int:
@@ -82,7 +81,7 @@ def delete_customer_confirmation(update: Update, context: CallbackContext) -> No
         success, response = customers_manager.delete_customer(customer_name)
         query.edit_message_text(text=response)
     elif query.data == 'cancel_delete':
-        query.edit_message_text(text='已取消删除操作。')
+        return cancel(update, context)
     return ConversationHandler.END
 
 """
@@ -93,16 +92,19 @@ def delete_customer_confirmation(update: Update, context: CallbackContext) -> No
 DELETE_PRODUCT = range(1)
 
 def start_delete_product(update: Update, context: CallbackContext) -> int:
-    product_list = inventory_manager.get_product_list()
-    keyboard = [[InlineKeyboardButton(product, callback_data=product)] for product in product_list]
-    reply_markup = InlineKeyboardMarkup(build_menu(keyboard, 2))
-    update.message.reply_text('请选择您要删除的产品名称：', reply_markup=reply_markup)
+    flag, reply_markup = get_product_buttons(update, context, manual_input=False, cancel=True) 
+    if flag is False:
+        return cancel(update, context)
+    message = '请选择您要删除的产品名称：'
+    reply_message(update, context, message, reply_markup)
     return DELETE_PRODUCT
 
 def delete_product(update: Update, context: CallbackContext) -> int:
     query = update.callback_query
     query.answer()
     selected_product = query.data
+    if selected_product.startswith('cancel_'):
+        return cancel(update, context)
     response = inventory_manager.delete_product_by_name(selected_product)
     response = f"已删除产品：{selected_product}"
     query.edit_message_text(text=response)
@@ -115,14 +117,11 @@ def delete_product(update: Update, context: CallbackContext) -> int:
 """
 (DELETE_ARRIVAL_DATE, DELETE_SUPPLIER, DELETE_PRODUCT, DELETE_SUPPLIER_CONFIRMATION) = range(4)
 MANUAL_DELETE_WAREHOUSING_DATE = range(1)
-def build_selection_keyboard(options):
-    # 根据选项列表生成选择键盘
-    keyboard = [InlineKeyboardButton(name, callback_data=name) for name in options]
-    keyboard.append(InlineKeyboardButton("取消", callback_data='cancel'))
-    return InlineKeyboardMarkup(build_menu(keyboard, 2))
 
 def start_delete_warehousing(update: Update, context: CallbackContext) -> int:
-    update.message.reply_text("请选择要删除的仓储记录的到货日期：", reply_markup=build_date_keyboard())
+    message = "请选择要删除的仓储记录的到货日期："
+    reply_markup=build_date_keyboard()
+    reply_message(update, context, message, reply_markup)
     return DELETE_ARRIVAL_DATE
 
 def delete_warehousing_arrival_date(update: Update, context: CallbackContext) -> int:
@@ -131,14 +130,13 @@ def delete_warehousing_arrival_date(update: Update, context: CallbackContext) ->
     if query.data == 'choose_today':
         selected_date = datetime.now().strftime('%d/%m/%Y')
         context.user_data['delete_arrival_date'] = selected_date
-        # supplier
-        flag, suppliers = warehousing_manager.get_supplier_list()  
+        flag, reply_markup = get_supplier_buttons(update, context, False, True)  
         if flag:
-            query.edit_message_text("请选择供应商名称：", reply_markup=build_selection_keyboard(suppliers))
+            query.edit_message_text("请选择供应商名称：", reply_markup=reply_markup)
             return DELETE_SUPPLIER
         else: 
-            query.edit_message_text("库存表中没有供应商名称，且库存表中无数据。\n无法删除库存数据。")
-            return ConversationHandler.END
+            query.edit_message_text("无法删除库存数据。")
+            return cancel(update, context)
     else: 
         query.edit_message_text(text="请输入日期：请使用DD/MM/YYYY格式输入")
         return MANUAL_DELETE_WAREHOUSING_DATE
@@ -150,14 +148,13 @@ def manual_delete_warehousing_date(update: Update, context: CallbackContext) -> 
         flag, custom_date = check_shipping_date(text)
         if flag:
             context.user_data['delete_arrival_date'] = custom_date
-            # supplier
-            flag, suppliers = warehousing_manager.get_supplier_list()  
+            flag, reply_markup = get_supplier_buttons(update, context, False, True)  
             if flag:
-                update.message.reply_text("请选择供应商名称：", reply_markup=build_selection_keyboard(suppliers))
+                update.message.reply_text("请选择供应商名称：", reply_markup=reply_markup)
                 return DELETE_SUPPLIER
             else: 
-                update.message.reply_text("库存表中没有供应商名称，且库存表中无数据。\n无法删除库存数据。")
-                return ConversationHandler.END
+                update.message.reply_text("无法删除库存数据。")
+                return cancel(update, context)
         else:
             update.message.reply_text("日期格式不正确，请使用DD/MM/YYYY格式重新输入：")
         return MANUAL_DELETE_WAREHOUSING_DATE
@@ -170,24 +167,20 @@ def delete_warehousing_supplier(update: Update, context: CallbackContext) -> int
     query = update.callback_query
     query.answer()
     if query.data == 'cancel':
-        query.edit_message_text("操作已取消。")
-        return ConversationHandler.END
+        return cancel(update, context)
     context.user_data['delete_supplier'] = query.data
-    flag, products = warehousing_manager.get_product_list() 
-    
+    flag, products = get_product_buttons(update, context, manual_input=False, cancel=True) 
     if flag:
-        query.edit_message_text("请选择产品名称：", reply_markup=build_selection_keyboard(products))
+        query.edit_message_text("请选择产品名称：", reply_markup=products)
         return DELETE_PRODUCT
     else: 
-        update.message.reply_text("库存表中没有产品名称，且库存表中无数据。\n无法删除库存数据。")
-        return ConversationHandler.END
+        return cancel(update, context)
     
 def delete_warehousing_product(update: Update, context: CallbackContext) -> int:
     query = update.callback_query
     query.answer()
     if query.data == 'cancel':
-        query.edit_message_text("操作已取消。")
-        return ConversationHandler.END
+        return cancel(update, context)
     context.user_data['delete_product'] = query.data
     confirmation_keyboard = InlineKeyboardMarkup([[InlineKeyboardButton("确认", callback_data='confirm_delete'), InlineKeyboardButton("取消", callback_data='cancel')]])
     query.edit_message_text(f"您要删除的记录信息如下：\n到货日期：{context.user_data['delete_arrival_date']}\n供应商：{context.user_data['delete_supplier']}\n产品名称：{context.user_data['delete_product']}\n请确认是否删除：", reply_markup=confirmation_keyboard)
